separate script for named entity "Person"
see #63
